+++ scripts/ci-env.ps1
@@ -8,13 +8,13 @@ switch ($name) {
         return "2019.4.39f1"
     }
     "unity2020" {
-        return "2020.3.38f1"
+        return "2020.3.39f1"
     }
     "unity2021" {
-        return "2021.3.8f1"
+        return "2021.3.10f1"
     }
     "unity2022" {
-        return "2022.1.14f1"
+        return "2022.1.16f1"
     }
     Default {
         throw "Unkown variable '$name'"