+++ test/Sentry.Unity.Tests/ContextWriterTests.cs
@@ -67,7 +67,7 @@ namespace Sentry.Unity.Tests
 
             };
             var context = new MockContextWriter();
-            var options = new SentryUnityOptions(_sentryMonoBehaviour, _testApplication, null, false)
+            var options = new SentryUnityOptions(_sentryMonoBehaviour, _testApplication, false)
             {
                 Dsn = "http://publickey@localhost/12345",
                 Enabled = true,